build: use lower version of line chart to be consistent with other nivo libraries
