+++ package.json
@@ -12,7 +12,7 @@
     "@mui/material": "^5.10.13",
     "@nivo/bar": "^0.80.0",
     "@nivo/core": "^0.80.0",
-    "@nivo/line": "^0.84.0",
+    "@nivo/line": "^0.80.0",
     "@nivo/pie": "^0.80.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
