+++ src/utils/ChartUtils.ts
@@ -34,37 +34,6 @@ export function convertToCommentsReceived(comments: Comment[]): ReviewBarChartSe
   return convertToItemsReceived(rawData);
 }
 
-export function convertToCommentsLineChart(comments: Comment[]) {
-  const data = comments.map((item) => {
-    const date = new Date(item.createdAt);
-
-    return {
-      id: item.reviewerName,
-      x: new Date(date.getFullYear(), date.getMonth()),
-    };
-  });
-
-  const grouped = tidy(data, groupBy(['id', 'x'], [summarize({ y: n() })]));
-
-  const transformedData = grouped.reduce((result, item) => {
-    const foundIndex = result.findIndex((i) => i.id === item.id);
-    if (foundIndex !== -1) {
-      result[foundIndex].data.push({ x: item.x, y: item.y });
-    } else {
-      result.push({ id: item.id, data: [{ x: item.x, y: item.y }] });
-    }
-    return result;
-  }, [] as Array<{ id: string; data: { x: Date; y: number }[] }>);
-
-  const filtered = transformedData.filter((item) => item.data.every((i) => i.y !== 0));
-
-  filtered.forEach((item) => {
-    item.data.sort((a, b) => a.x.getTime() - b.x.getTime());
-  });
-
-  return filtered;
-}
-
 export function convertToFilesCommented(comments: Comment[]): ReviewBarChartSettings {
   const paths = comments.filter((item) => !!item.filePath).map((item) => ({ filePath: item.filePath }));
   const authors = tidy(comments, distinct(['reviewerName'])).map((item) => item.reviewerName);
